pull out custom adapters from AmqApi
